Declaration of const table
I don't know if it works properly, but know it compile. I changed the
declaration of _RS_acc_mean_table.
Now it's declared in RS_accelerometer_utility.h as a constant table. So
I deleted the .s file.

We have to modify the declaration and the controler setup to work with
fixed point arithmetic (multipling by 1/lengthin the MAC instruction)